+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix the bug in the PDC.predict() output
+
 ## [0.2.0] - 2024-07-15
 ### Added
 - Add weighting method and their benchmark